+++ packages/ui-vue/src/input/index.ts
@@ -34,3 +34,20 @@ export type {
   ZRadioValue,
   ZRadioOption,
 } from './ZRadioGroup.vue'
+
+export { default as ZSelect } from './ZSelect.vue'
+export type {
+  ZSelectProps,
+  ZSelectEmits,
+  ZSelectValue,
+  ZSelectOption,
+  ZSelectSize,
+} from './ZSelect.vue'
+
+export { default as ZForm } from './ZForm.vue'
+export type { ZFormProps, ZFormExpose } from './ZForm.vue'
+
+export { default as ZFormItem } from './ZFormItem.vue'
+export type { ZFormItemProps } from './ZFormItem.vue'
+
+export type { ZFormRule, ZFormRules } from './_form-ctx'